backend/calorieModel/requirements.txt
torch==2.7.1
torchvision==0.22.1
pillow==11.2.1
flask==3.1.0
flask-cors==5.0.0